Repackage migration, separate inner classes
