+++ src/main/java/me/lokka30/treasury/plugin/command/treasury/subcommand/migration/PhasedSubscriber.java
@@ -1,0 +1,33 @@
+package me.lokka30.treasury.plugin.command.treasury.subcommand.migration;
+
+import me.lokka30.treasury.api.economy.response.EconomyException;
+import me.lokka30.treasury.api.economy.response.EconomySubscriber;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.concurrent.Phaser;
+
+abstract class PhasedSubscriber<T> implements EconomySubscriber<T> {
+    private final Phaser phaser;
+
+    PhasedSubscriber(@NotNull Phaser phaser) {
+        this.phaser = phaser;
+        this.phaser.register();
+    }
+
+    @Override
+    public final void succeed(@NotNull T t) {
+        phaseAccept(t);
+        phaser.arriveAndDeregister();
+    }
+
+    public abstract void phaseAccept(@NotNull T t);
+
+    @Override
+    public final void fail(@NotNull EconomyException exception) {
+        phaseFail(exception);
+        phaser.arriveAndDeregister();
+    }
+
+    public abstract void phaseFail(@NotNull EconomyException exception);
+
+}
